+++ main/display_alert_adapter.h
@@ -1,6 +1,8 @@
 #ifndef DISPLAY_ALERT_ADAPTER_H
 #define DISPLAY_ALERT_ADAPTER_H
 
+#include <stdbool.h>
+
 #include "esp_err.h"
 
 #ifdef __cplusplus
@@ -10,6 +12,7 @@ extern "C" {
 esp_err_t display_alert_adapter_init(void);
 esp_err_t display_alert_adapter_show_danger_overlay(void);
 esp_err_t display_alert_adapter_hide_danger_overlay(void);
+esp_err_t display_alert_adapter_set_suppressed(bool suppressed);
 void display_alert_adapter_process_ui(void);
 
 #ifdef __cplusplus